+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Changed signature of `init` event, it now has `postId` parameter. See docs at [GitHub](https://github.com/bologer/anycomment.io/wiki/Client-side)
 - It is now possible to renders comments for specific page by providing `pageId` parameter, #466
 - Now loading default font Google font in async manner, #257
+- Added ability to launch comments programmatically using new manager API, see docs at [GitHub](https://github.com/bologer/anycomment.io/wiki/Client-side#programmatically-initiate-comments), #476
 
 **Fixes:**